requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pdfplumber
tiktoken
chromadb
sentence-transformers
httpx
numpy
python-dotenv
